chore(runner): increase timeout in serializeReadableStream test
This commit increases the timeout in one of the tests for the `serializeReadableStream` function. The timeout in the "should emit the 'data' event when sending the 'read' event" test has been increased from 1ms to 10ms. This change aims to improve the test's reliability by allowing more time for the asynchronous operations to complete, reducing the likelihood of false negatives due to timing issues.