feat(api)!: remove project variables and secrets functionality
Removed the project variables and secrets functionality from the project module as it will be moved to a dedicated module. This change includes the removal of related entities (ProjectSecret, ProjectVariable), their corresponding routes, and utility functions for encryption/decryption. This is a breaking change as it removes several API endpoints and database entities that were previously available in the project module.

The functionality will be reimplemented in a new dedicated module to better separate concerns and provide more focused management of project configuration values.