chore(api): update Thread entity events property type
Updates the type definition of the `events` property in the Thread entity to include `null` as a possible value. This change better reflects the actual behavior of the property, which can be null when a thread is first created or when events haven't been loaded yet, improving type safety throughout the application.